Add report stream for `CAMPAIGN` level
