+++ airbyte-integrations/connectors/source-pinterest/Dockerfile
@@ -34,5 +34,5 @@ COPY source_pinterest ./source_pinterest
 ENV AIRBYTE_ENTRYPOINT "python /airbyte/integration_code/main.py"
 ENTRYPOINT ["python", "/airbyte/integration_code/main.py"]
 
-LABEL io.airbyte.version=0.5.3
+LABEL io.airbyte.version=0.6.0
 LABEL io.airbyte.name=airbyte/source-pinterest
