+++ airbyte-integrations/connectors/source-pinterest/integration_tests/abnormal_state.json
@@ -86,5 +86,16 @@
         "name": "ad_analytics"
       }
     }
+  },
+  {
+    "type": "STREAM",
+    "stream": {
+      "stream_state": {
+        "DATE": "3021-06-09"
+      },
+      "stream_descriptor": {
+        "name": "campaign_analytics_report"
+      }
+    }
   }
 ]
