+++ airbyte-integrations/connectors/source-pinterest/integration_tests/expected_records.jsonl
@@ -1,3 +1,4 @@
 {"stream": "ad_accounts", "data": {"id": "549761668032", "name": "Airbyte", "owner": {"username": "integrationtest0375", "id": "666744057242074926"}, "country": "US", "currency": "USD", "permissions": ["OWNER"], "created_time": 1603772920, "updated_time": 1623173784}, "emitted_at": 1688461289470}
 {"stream": "boards", "data": {"media": {"pin_thumbnail_urls": [], "image_cover_url": "https://i.pinimg.com/400x300/c6/b6/0d/c6b60d6b5f2ec04db7748d35fb1a8004.jpg"}, "owner": {"username": "integrationtest0375"}, "created_at": "2021-06-08T09:37:18", "board_pins_modified_at": "2021-10-25T11:17:56.715000", "id": "666743988523388559", "collaborator_count": 0, "follower_count": 2, "pin_count": 1, "privacy": "PUBLIC", "name": "business", "description": ""}, "emitted_at": 1680356853019}
 {"stream":"board_pins","data":{"link":"http://airbyte.io/","dominant_color":"#cacafe","media":{"media_type":"image","images":{"150x150":{"width":150,"height":150,"url":"https://i.pinimg.com/150x150/c6/b6/0d/c6b60d6b5f2ec04db7748d35fb1a8004.jpg"},"400x300":{"width":400,"height":300,"url":"https://i.pinimg.com/400x300/c6/b6/0d/c6b60d6b5f2ec04db7748d35fb1a8004.jpg"},"600x":{"width":600,"height":359,"url":"https://i.pinimg.com/600x/c6/b6/0d/c6b60d6b5f2ec04db7748d35fb1a8004.jpg"},"1200x":{"width":1200,"height":718,"url":"https://i.pinimg.com/1200x/c6/b6/0d/c6b60d6b5f2ec04db7748d35fb1a8004.jpg"}}},"is_standard":true,"creative_type":"REGULAR","is_owner":true,"board_section_id":"5195034916661798218","id":"666743919837294988","description":"Data Integration","has_been_promoted":true,"created_at":"2021-06-08T09:37:30","note":"","product_tags":[],"alt_text":null,"title":"Airbyte","board_owner":{"username":"integrationtest0375"},"parent_pin_id":null,"board_id":"666743988523388559"},"emitted_at":1688054568572}
+{"stream": "campaign_analytics_report", "data": {"ADVERTISER_ID": 549761668032.0, "AD_ACCOUNT_ID": "549761668032", "CAMPAIGN_DAILY_SPEND_CAP": 750000.0, "CAMPAIGN_ENTITY_STATUS": "ACTIVE", "CAMPAIGN_ID": 626744128982.0, "CAMPAIGN_LIFETIME_SPEND_CAP": 0.0, "CAMPAIGN_NAME": "2021-06-08 09:08 UTC | Brand awareness", "IMPRESSION_2": 3.0, "TOTAL_IMPRESSION_FREQUENCY": 1.5, "TOTAL_IMPRESSION_USER": 2.0, "DATE": "2023-07-14"}, "emitted_at": 1690299367301}
