+++ airbyte-integrations/connectors/source-pinterest/metadata.yaml
@@ -5,7 +5,7 @@ data:
   connectorSubtype: api
   connectorType: source
   definitionId: 5cb7e5fe-38c2-11ec-8d3d-0242ac130003
-  dockerImageTag: 0.5.3
+  dockerImageTag: 0.6.0
   dockerRepository: airbyte/source-pinterest
   githubIssueLabel: source-pinterest
   icon: pinterest.svg
