+++ airbyte-integrations/connectors/source-pinterest/source_pinterest/reports/__init__.py
@@ -1,0 +1,7 @@
+#
+# Copyright (c) 2021 Airbyte, Inc., all rights reserved.
+#
+
+from .reports import CampaignAnalyticsReport
+
+__all__ = ["CampaignAnalyticsReport"]
