+++ airbyte-integrations/connectors/source-pinterest/unit_tests/conftest.py
@@ -5,6 +5,7 @@
 from unittest.mock import MagicMock
 
 from pytest import fixture
+from source_pinterest.reports import CampaignAnalyticsReport
 
 
 @fixture
@@ -62,3 +63,17 @@ def test_response_filter(test_record_filter):
     response = MagicMock()
     response.json.return_value = test_record_filter
     return response
+
+
+@fixture
+def analytics_report_stream():
+    return CampaignAnalyticsReport(parent=None, config=MagicMock())
+
+
+@fixture
+def date_range():
+    return {
+        'start_date': '2023-01-01',
+        'end_date': '2023-01-31',
+        'parent': {'id': '123'}
+    }
