+++ airbyte-integrations/connectors/source-pinterest/unit_tests/test_incremental_streams.py
@@ -9,7 +9,7 @@ from unittest.mock import MagicMock
 import pytest
 from airbyte_cdk.models import SyncMode
 from pytest import fixture
-from source_pinterest.source import AdAccountAnalytics, Campaigns, IncrementalPinterestSubStream
+from source_pinterest.streams import AdAccountAnalytics, Campaigns, IncrementalPinterestSubStream
 
 
 @fixture
